Moved vector series into a subfolder.
