+++ js/Series/Vector/VectorSeries.js
@@ -23,13 +23,12 @@ var __extends = (this && this.__extends) || (function () {
         d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
     };
 })();
-import A from '../Core/Animation/AnimationUtilities.js';
+import A from '../../Core/Animation/AnimationUtilities.js';
 var animObject = A.animObject;
-import BaseSeries from '../Core/Series/Series.js';
-var ScatterSeries = BaseSeries.seriesTypes.scatter;
-import H from '../Core/Globals.js';
-import LineSeries from './Line/LineSeries.js';
-import U from '../Core/Utilities.js';
+import BaseSeries from '../../Core/Series/Series.js';
+var _a = BaseSeries.seriesTypes, LineSeries = _a.line, ScatterSeries = _a.scatter;
+import H from '../../Core/Globals.js';
+import U from '../../Core/Utilities.js';
 var arrayMax = U.arrayMax, extend = U.extend, merge = U.merge, pick = U.pick;
 /* *
  *
